# Run OCI image package building and publishing concurrently