--- a/cmd/plugin/builder/command/cli_compile.go
+++ b/cmd/plugin/builder/command/cli_compile.go
@@ -72,30 +72,6 @@ type PluginCompileArgs struct {
 
 const local = "local"
 
-var identifiers = []string{
-	string('\U0001F435'),
-	string('\U0001F43C'),
-	string('\U0001F436'),
-	string('\U0001F430'),
-	string('\U0001F98A'),
-	string('\U0001F431'),
-	string('\U0001F981'),
-	string('\U0001F42F'),
-	string('\U0001F42E'),
-	string('\U0001F437'),
-	string('\U0001F42D'),
-	string('\U0001F428'),
-}
-
-func getID(i int) string {
-	index := i
-	if i >= len(identifiers) {
-		// Well aren't you lucky
-		index = i % len(identifiers)
-	}
-	return identifiers[index]
-}
-
 func getBuildArch(arch []string) []cli.Arch {
 	var arrArch []cli.Arch
 	for _, buildArch := range arch {
@@ -110,12 +86,6 @@ func getBuildArch(arch []string) []cli.Arch {
 		}
 	}
 	return arrArch
-}
-
-type errInfo struct {
-	Err  error
-	Path string
-	ID   string
 }
 
 // setGlobals initializes a set of global variables used throughout the compile
@@ -155,10 +125,10 @@ func Compile(compileArgs *PluginCompileArgs) error {
 	guard := make(chan struct{}, maxConcurrent)
 
 	// Mix up IDs so we don't always get the same set.
-	randSkew := rand.Intn(len(identifiers)) // nolint:gosec
+	randSkew := rand.Intn(len(helpers.Identifiers)) // nolint:gosec
 	var wg sync.WaitGroup
 	plugins := make(chan cli.Plugin, len(files))
-	fatalErrors := make(chan errInfo, len(files))
+	fatalErrors := make(chan helpers.ErrInfo, len(files))
 	g := glob.MustCompile(compileArgs.Match)
 	for i, f := range files {
 		if f.IsDir() {
@@ -169,7 +139,7 @@ func Compile(compileArgs *PluginCompileArgs) error {
 					defer wg.Done()
 					p, err := buildPlugin(fullPath, id)
 					if err != nil {
-						fatalErrors <- errInfo{Err: err, Path: fullPath, ID: id}
+						fatalErrors <- helpers.ErrInfo{Err: err, Path: fullPath, ID: id}
 					} else {
 						plug := cli.Plugin{
 							Name:        p.Name,
@@ -180,7 +150,7 @@ func Compile(compileArgs *PluginCompileArgs) error {
 						plugins <- plug
 					}
 					<-guard
-				}(filepath.Join(compileArgs.SourcePath, f.Name()), getID(i+randSkew))
+				}(filepath.Join(compileArgs.SourcePath, f.Name()), helpers.GetID(i+randSkew))
 			}
 		}
 	}