--- a/cmd/plugin/builder/plugin/build-packages.go
+++ b/cmd/plugin/builder/plugin/build-packages.go
@@ -6,10 +6,13 @@ package plugin
 
 import (
 	"fmt"
+	"math/rand"
 	"os"
 	"path/filepath"
+	"sync"
 
 	"github.com/pkg/errors"
+	kerrors "k8s.io/apimachinery/pkg/util/errors"
 
 	"github.com/vmware-tanzu/tanzu-plugin-runtime/log"
 
@@ -47,48 +50,79 @@ func (bpo *BuildPluginPackageOptions) BuildPluginPackages() error {
 
 	log.Infof("Using plugin binary artifacts from %q", bpo.BinaryArtifactDir)
 
+	// Limit the number of concurrent operations we perform so we don't overwhelm the system.
+	maxConcurrent := helpers.GetMaxParallelism()
+	guard := make(chan struct{}, maxConcurrent)
+	var wg sync.WaitGroup
+	// Mix up IDs so we don't always get the same set.
+	randSkew := rand.Intn(len(helpers.Identifiers)) // nolint:gosec
+	errorList := []error{}
+
+	generatePluginPackage := func(p cli.Plugin, osArch cli.Arch, version, id string) {
+		defer func() {
+			<-guard
+			wg.Done()
+		}()
+
+		pluginBinaryFilePath := filepath.Join(bpo.BinaryArtifactDir, osArch.OS(), osArch.Arch(),
+			p.Target, p.Name, version,
+			cli.MakeArtifactName(p.Name, osArch))
+
+		if !utils.PathExists(pluginBinaryFilePath) {
+			return
+		}
+
+		// Check whether the binary exec file is valid
+		valid, err := helpers.ValidatePluginBinary(pluginBinaryFilePath)
+
+		// Return err if plugin binary file validation fails
+		if err != nil {
+			errorList = append(errorList, err)
+			return
+		}
+
+		// Return err if plugin binary file is not valid
+		if !valid {
+			errorList = append(errorList, fmt.Errorf("invalid plugin binary :%v", pluginBinaryFilePath))
+			return
+		}
+
+		pluginTarFilePath := filepath.Join(bpo.PackageArtifactDir, helpers.GetPluginArchiveRelativePath(p, osArch, version))
+		image := fmt.Sprintf("%s/plugins/%s/%s/%s:%s", bpo.LocalOCIRegistry, osArch.OS(), osArch.Arch(), p.Name, version)
+
+		log.Infof("Generating plugin package for 'plugin:%s' 'target:%s' 'os:%s' 'arch:%s' 'version:%s'", p.Name, p.Target, osArch.OS(), osArch.Arch(), version)
+
+		err = carvelhelpers.NewImageOperationsImpl().PushImage(image, []string{pluginBinaryFilePath})
+		if err != nil {
+			errorList = append(errorList, errors.Wrapf(err, "unable to push package to temporary registry for plugin: %s, target: %s, os: %s, arch: %s, version: %s", p.Name, p.Target, osArch.OS(), osArch.Arch(), version))
+			return
+		}
+
+		err = bpo.CraneOptions.SaveImage(image, pluginTarFilePath)
+		if err != nil {
+			errorList = append(errorList, errors.Wrapf(err, "unable to generate package for plugin: %s, target: %s, os: %s, arch: %s, version: %s", p.Name, p.Target, osArch.OS(), osArch.Arch(), version))
+			return
+		}
+
+		log.Infof("Generated plugin package at %q", pluginTarFilePath)
+	}
+
+	id := 0
 	for i := range pluginManifest.Plugins {
 		for _, osArch := range cli.AllOSArch {
 			for _, version := range pluginManifest.Plugins[i].Versions {
-				pluginBinaryFilePath := filepath.Join(bpo.BinaryArtifactDir, osArch.OS(), osArch.Arch(),
-					pluginManifest.Plugins[i].Target, pluginManifest.Plugins[i].Name, version,
-					cli.MakeArtifactName(pluginManifest.Plugins[i].Name, osArch))
-
-				if !utils.PathExists(pluginBinaryFilePath) {
-					continue
-				}
-
-				// Check whether the binary exec file is valid
-				valid, err := helpers.ValidatePluginBinary(pluginBinaryFilePath)
-
-				// Return err if plugin binary file validation fails
-				if err != nil {
-					return err
-				}
-
-				// Return err if plugin binary file is not valid
-				if !valid {
-					return fmt.Errorf("invalid plugin binary :%v", pluginBinaryFilePath)
-				}
-
-				pluginTarFilePath := filepath.Join(bpo.PackageArtifactDir, helpers.GetPluginArchiveRelativePath(pluginManifest.Plugins[i], osArch, version))
-				image := fmt.Sprintf("%s/plugins/%s/%s/%s:%s", bpo.LocalOCIRegistry, osArch.OS(), osArch.Arch(), pluginManifest.Plugins[i].Name, version)
-
-				log.Infof("Generating plugin package for 'plugin:%s' 'target:%s' 'os:%s' 'arch:%s' 'version:%s'", pluginManifest.Plugins[i].Name, pluginManifest.Plugins[i].Target, osArch.OS(), osArch.Arch(), version)
-
-				err = carvelhelpers.NewImageOperationsImpl().PushImage(image, []string{pluginBinaryFilePath})
-				if err != nil {
-					return errors.Wrapf(err, "unable to push package to temporary registry for plugin: %s, target: %s, os: %s, arch: %s, version: %s", pluginManifest.Plugins[i].Name, pluginManifest.Plugins[i].Target, osArch.OS(), osArch.Arch(), version)
-				}
-
-				err = bpo.CraneOptions.SaveImage(image, pluginTarFilePath)
-				if err != nil {
-					return errors.Wrapf(err, "unable to generate package for plugin: %s, target: %s, os: %s, arch: %s, version: %s", pluginManifest.Plugins[i].Name, pluginManifest.Plugins[i].Target, osArch.OS(), osArch.Arch(), version)
-				}
-
-				log.Infof("Generated plugin package at %q", pluginTarFilePath)
+				wg.Add(1)
+				guard <- struct{}{}
+				go generatePluginPackage(pluginManifest.Plugins[i], osArch, version, helpers.GetID(id+randSkew))
+				id++
 			}
 		}
+	}
+
+	wg.Wait()
+
+	if len(errorList) != 0 {
+		return kerrors.NewAggregate(errorList)
 	}
 
 	// copy plugin_manifest.yaml to PackageArtifactDir