--- a/cmd/plugin/builder/plugin/build-packages.go
+++ b/cmd/plugin/builder/plugin/build-packages.go
@@ -6,13 +6,11 @@ package plugin
 
 import (
 	"fmt"
-	"math/rand"
 	"os"
 	"path/filepath"
 	"sync"
 
 	"github.com/pkg/errors"
-	kerrors "k8s.io/apimachinery/pkg/util/errors"
 
 	"github.com/vmware-tanzu/tanzu-plugin-runtime/log"
 
@@ -54,11 +52,9 @@ func (bpo *BuildPluginPackageOptions) BuildPluginPackages() error {
 	maxConcurrent := helpers.GetMaxParallelism()
 	guard := make(chan struct{}, maxConcurrent)
 	var wg sync.WaitGroup
-	// Mix up IDs so we don't always get the same set.
-	randSkew := rand.Intn(len(helpers.Identifiers)) // nolint:gosec
-	errorList := []error{}
+	fatalErrors := make(chan helpers.ErrInfo, len(pluginManifest.Plugins)*len(cli.AllOSArch))
 
-	generatePluginPackage := func(p cli.Plugin, osArch cli.Arch, version, id string) {
+	generatePluginPackage := func(p cli.Plugin, osArch cli.Arch, version, threadID string) {
 		defer func() {
 			<-guard
 			wg.Done()
@@ -68,43 +64,10 @@ func (bpo *BuildPluginPackageOptions) BuildPluginPackages() error {
 			p.Target, p.Name, version,
 			cli.MakeArtifactName(p.Name, osArch))
 
-		if !utils.PathExists(pluginBinaryFilePath) {
-			return
-		}
-
-		// Check whether the binary exec file is valid
-		valid, err := helpers.ValidatePluginBinary(pluginBinaryFilePath)
-
-		// Return err if plugin binary file validation fails
+		err := bpo.generatePluginPackage(pluginBinaryFilePath, p, osArch, version, threadID)
 		if err != nil {
-			errorList = append(errorList, err)
-			return
+			fatalErrors <- helpers.ErrInfo{Err: err, ID: threadID, Path: pluginBinaryFilePath}
 		}
-
-		// Return err if plugin binary file is not valid
-		if !valid {
-			errorList = append(errorList, fmt.Errorf("invalid plugin binary :%v", pluginBinaryFilePath))
-			return
-		}
-
-		pluginTarFilePath := filepath.Join(bpo.PackageArtifactDir, helpers.GetPluginArchiveRelativePath(p, osArch, version))
-		image := fmt.Sprintf("%s/plugins/%s/%s/%s:%s", bpo.LocalOCIRegistry, osArch.OS(), osArch.Arch(), p.Name, version)
-
-		log.Infof("Generating plugin package for 'plugin:%s' 'target:%s' 'os:%s' 'arch:%s' 'version:%s'", p.Name, p.Target, osArch.OS(), osArch.Arch(), version)
-
-		err = carvelhelpers.NewImageOperationsImpl().PushImage(image, []string{pluginBinaryFilePath})
-		if err != nil {
-			errorList = append(errorList, errors.Wrapf(err, "unable to push package to temporary registry for plugin: %s, target: %s, os: %s, arch: %s, version: %s", p.Name, p.Target, osArch.OS(), osArch.Arch(), version))
-			return
-		}
-
-		err = bpo.CraneOptions.SaveImage(image, pluginTarFilePath)
-		if err != nil {
-			errorList = append(errorList, errors.Wrapf(err, "unable to generate package for plugin: %s, target: %s, os: %s, arch: %s, version: %s", p.Name, p.Target, osArch.OS(), osArch.Arch(), version))
-			return
-		}
-
-		log.Infof("Generated plugin package at %q", pluginTarFilePath)
 	}
 
 	id := 0
@@ -113,16 +76,22 @@ func (bpo *BuildPluginPackageOptions) BuildPluginPackages() error {
 			for _, version := range pluginManifest.Plugins[i].Versions {
 				wg.Add(1)
 				guard <- struct{}{}
-				go generatePluginPackage(pluginManifest.Plugins[i], osArch, version, helpers.GetID(id+randSkew))
+				go generatePluginPackage(pluginManifest.Plugins[i], osArch, version, helpers.GetID(id))
 				id++
 			}
 		}
 	}
 
 	wg.Wait()
+	close(fatalErrors)
 
-	if len(errorList) != 0 {
-		return kerrors.NewAggregate(errorList)
+	hasFailed := false
+	for err := range fatalErrors {
+		hasFailed = true
+		log.Errorf("%s - building plugin package for %q failed - %v", err.ID, err.Path, err.Err)
+	}
+	if hasFailed {
+		os.Exit(1)
 	}
 
 	// copy plugin_manifest.yaml to PackageArtifactDir
@@ -132,5 +101,42 @@ func (bpo *BuildPluginPackageOptions) BuildPluginPackages() error {
 	}
 	log.Infof("Saved plugin manifest at %q", filepath.Join(bpo.PackageArtifactDir, cli.PluginManifestFileName))
 
+	return nil
+}
+
+func (bpo *BuildPluginPackageOptions) generatePluginPackage(pluginBinaryFilePath string, p cli.Plugin, osArch cli.Arch, version, threadID string) error {
+	if !utils.PathExists(pluginBinaryFilePath) {
+		return nil
+	}
+
+	// Check whether the binary exec file is valid
+	valid, err := helpers.ValidatePluginBinary(pluginBinaryFilePath, threadID)
+
+	// Return err if plugin binary file validation fails
+	if err != nil {
+		return err
+	}
+
+	// Return err if plugin binary file is not valid
+	if !valid {
+		return fmt.Errorf("invalid plugin binary :%v", pluginBinaryFilePath)
+	}
+
+	pluginTarFilePath := filepath.Join(bpo.PackageArtifactDir, helpers.GetPluginArchiveRelativePath(p, osArch, version))
+	image := fmt.Sprintf("%s/plugins/%s/%s/%s:%s", bpo.LocalOCIRegistry, osArch.OS(), osArch.Arch(), p.Name, version)
+
+	log.Infof("%s Generating plugin package for 'plugin:%s' 'target:%s' 'os:%s' 'arch:%s' 'version:%s'", threadID, p.Name, p.Target, osArch.OS(), osArch.Arch(), version)
+
+	err = carvelhelpers.NewImageOperationsImpl().PushImage(image, []string{pluginBinaryFilePath})
+	if err != nil {
+		return errors.Wrapf(err, "unable to push package to temporary registry for plugin: %s, target: %s, os: %s, arch: %s, version: %s", p.Name, p.Target, osArch.OS(), osArch.Arch(), version)
+	}
+
+	err = bpo.CraneOptions.SaveImage(image, pluginTarFilePath)
+	if err != nil {
+		return errors.Wrapf(err, "unable to generate package for plugin: %s, target: %s, os: %s, arch: %s, version: %s", p.Name, p.Target, osArch.OS(), osArch.Arch(), version)
+	}
+
+	log.Infof("%s Generated plugin package at %q", threadID, pluginTarFilePath)
 	return nil
 }