--- a/cmd/plugin/builder/plugin/build-packages.go
+++ b/cmd/plugin/builder/plugin/build-packages.go
@@ -52,7 +52,7 @@ func (bpo *BuildPluginPackageOptions) BuildPluginPackages() error {
 	maxConcurrent := helpers.GetMaxParallelism()
 	guard := make(chan struct{}, maxConcurrent)
 	var wg sync.WaitGroup
-	fatalErrors := make(chan helpers.ErrInfo, len(pluginManifest.Plugins)*len(cli.AllOSArch))
+	fatalErrors := make(chan helpers.ErrInfo, helpers.GetNumberOfIndividualPluginBinariesFromManifest(pluginManifest))
 
 	generatePluginPackage := func(p cli.Plugin, osArch cli.Arch, version, threadID string) {
 		defer func() {