--- a/cmd/plugin/builder/plugin/publish-packages.go
+++ b/cmd/plugin/builder/plugin/publish-packages.go
@@ -5,12 +5,11 @@ package plugin
 
 import (
 	"fmt"
-	"math/rand"
+	"os"
 	"path/filepath"
 	"sync"
 
 	"github.com/pkg/errors"
-	kerrors "k8s.io/apimachinery/pkg/util/errors"
 
 	"github.com/vmware-tanzu/tanzu-plugin-runtime/log"
 
@@ -50,33 +49,19 @@ func (ppo *PublishPluginPackageOptions) PublishPluginPackages() error {
 	maxConcurrent := helpers.GetMaxParallelism()
 	guard := make(chan struct{}, maxConcurrent)
 	var wg sync.WaitGroup
-	// Mix up IDs so we don't always get the same set.
-	randSkew := rand.Intn(len(helpers.Identifiers)) // nolint:gosec
-	errorList := []error{}
+	fatalErrors := make(chan helpers.ErrInfo, len(pluginManifest.Plugins)*len(cli.AllOSArch))
 
-	publishPluginPackage := func(p cli.Plugin, osArch cli.Arch, version, id string) {
+	publishPluginPackage := func(p cli.Plugin, osArch cli.Arch, version, threadID string) {
 		defer func() {
 			<-guard
 			wg.Done()
 		}()
 
 		pluginTarFilePath := filepath.Join(ppo.PackageArtifactDir, helpers.GetPluginArchiveRelativePath(p, osArch, version))
-		if !utils.PathExists(pluginTarFilePath) {
-			return
-		}
 
-		imageToPush := fmt.Sprintf("%s/%s/%s/%s/%s/%s/%s:%s", ppo.Repository, ppo.Vendor, ppo.Publisher, osArch.OS(), osArch.Arch(), p.Target, p.Name, version)
-
-		if ppo.DryRun {
-			log.Infof("command: 'crane push %s %s'", pluginTarFilePath, imageToPush)
-		} else {
-			log.Infof("publishing plugin 'name:%s' 'target:%s' 'os:%s' 'arch:%s' 'version:%s'", p.Name, p.Target, osArch.OS(), osArch.Arch(), version)
-			err = ppo.CraneOptions.PushImage(pluginTarFilePath, imageToPush)
-			if err != nil {
-				errorList = append(errorList, errors.Wrapf(err, "unable to publish plugin (name:%s, target:%s, os:%s, arch:%s, version:%s)", p.Name, p.Target, osArch.OS(), osArch.Arch(), version))
-				return
-			}
-			log.Infof("published plugin at '%s'", imageToPush)
+		err = ppo.publishPluginPackage(pluginTarFilePath, p, osArch, version, threadID)
+		if err != nil {
+			fatalErrors <- helpers.ErrInfo{Err: err, ID: threadID, Path: pluginTarFilePath}
 		}
 	}
 
@@ -86,7 +71,7 @@ func (ppo *PublishPluginPackageOptions) PublishPluginPackages() error {
 			for _, version := range pluginManifest.Plugins[i].Versions {
 				wg.Add(1)
 				guard <- struct{}{}
-				go publishPluginPackage(pluginManifest.Plugins[i], osArch, version, helpers.GetID(id+randSkew))
+				go publishPluginPackage(pluginManifest.Plugins[i], osArch, version, helpers.GetID(id))
 				id++
 			}
 		}
@@ -94,6 +79,36 @@ func (ppo *PublishPluginPackageOptions) PublishPluginPackages() error {
 
 	// wait for all WaitGroup to complete before continuing
 	wg.Wait()
+	close(fatalErrors)
 
-	return kerrors.NewAggregate(errorList)
+	hasFailed := false
+	for err := range fatalErrors {
+		hasFailed = true
+		log.Errorf("%s - publishing plugin package for %q failed - %v", err.ID, err.Path, err.Err)
+	}
+	if hasFailed {
+		os.Exit(1)
+	}
+
+	return nil
+}
+
+func (ppo *PublishPluginPackageOptions) publishPluginPackage(pluginTarFilePath string, p cli.Plugin, osArch cli.Arch, version, threadID string) error {
+	if !utils.PathExists(pluginTarFilePath) {
+		return nil
+	}
+
+	imageToPush := fmt.Sprintf("%s/%s/%s/%s/%s/%s/%s:%s", ppo.Repository, ppo.Vendor, ppo.Publisher, osArch.OS(), osArch.Arch(), p.Target, p.Name, version)
+
+	if ppo.DryRun {
+		log.Infof("%s command: 'crane push %s %s'", threadID, pluginTarFilePath, imageToPush)
+	} else {
+		log.Infof("%s publishing plugin 'name:%s' 'target:%s' 'os:%s' 'arch:%s' 'version:%s'", threadID, p.Name, p.Target, osArch.OS(), osArch.Arch(), version)
+		err := ppo.CraneOptions.PushImage(pluginTarFilePath, imageToPush)
+		if err != nil {
+			return errors.Wrapf(err, "unable to publish plugin (name:%s, target:%s, os:%s, arch:%s, version:%s)", p.Name, p.Target, osArch.OS(), osArch.Arch(), version)
+		}
+		log.Infof("%s published plugin at '%s'", threadID, imageToPush)
+	}
+	return nil
 }