… cmd/plugin/builder/helpers/helpers_test.go